# Vectorization of boolean arrays