--- a/ipa-core/src/ff/boolean.rs
+++ b/ipa-core/src/ff/boolean.rs
@@ -215,10 +215,20 @@ impl FieldVectorizable<256> for Boolean {
 #[cfg(all(test, unit_test))]
 mod test {
     use generic_array::GenericArray;
+    use proptest::prelude::{prop, Arbitrary, Strategy};
     use rand::{thread_rng, Rng};
     use typenum::U1;
 
     use crate::ff::{boolean::Boolean, Serializable};
+
+    impl Arbitrary for Boolean {
+        type Parameters = <bool as Arbitrary>::Parameters;
+        type Strategy = prop::strategy::Map<<bool as Arbitrary>::Strategy, fn(bool) -> Self>;
+
+        fn arbitrary_with(args: Self::Parameters) -> Self::Strategy {
+            <bool>::arbitrary_with(args).prop_map(Boolean)
+        }
+    }
 
     ///test serialize and deserialize
     #[test]