--- a/ipa-core/src/ff/mod.rs
+++ b/ipa-core/src/ff/mod.rs
@@ -11,6 +11,7 @@ mod galois_field;
 mod prime_field;
 
 use std::{
+    borrow::Borrow,
     convert::Infallible,
     ops::{Add, AddAssign, Sub, SubAssign},
 };
@@ -93,6 +94,24 @@ pub trait ArrayAccess {
     fn iter(&self) -> Self::Iter<'_>;
 }
 
+pub trait ArrayAccessRef {
+    type Element;
+    type Ref<'a>: Borrow<Self::Element> + Clone
+    where
+        Self: 'a;
+    type Iter<'a>: Iterator<Item = Self::Ref<'a>> + ExactSizeIterator + Send
+    where
+        Self: 'a;
+
+    fn get(&self, index: usize) -> Option<Self::Ref<'_>>;
+
+    fn set(&mut self, index: usize, e: Self::Ref<'_>);
+
+    fn iter(&self) -> Self::Iter<'_>;
+
+    fn make_ref(src: &Self::Element) -> Self::Ref<'_>;
+}
+
 pub trait Expand {
     type Input;
 
@@ -105,7 +124,9 @@ pub trait Expand {
 /// supports `FromIterator` to collect an iterator of elements back into the original type
 pub trait CustomArray
 where
-    Self: ArrayAccess<Output = Self::Element> + Expand<Input = Self::Element>,
+    Self: ArrayAccess<Output = Self::Element>
+        + Expand<Input = Self::Element>
+        + ArrayBuild<Input = Self::Element>,
 {
     type Element;
 }
@@ -113,7 +134,30 @@ where
 /// impl Custom Array for all compatible structs
 impl<S> CustomArray for S
 where
-    S: ArrayAccess + Expand<Input = <S as ArrayAccess>::Output>,
+    S: ArrayAccess
+        + Expand<Input = <S as ArrayAccess>::Output>
+        + ArrayBuild<Input = <S as ArrayAccess>::Output>,
 {
     type Element = <S as ArrayAccess>::Output;
+}
+
+pub trait ArrayBuild {
+    type Input;
+    type Builder: ArrayBuilder<Element = Self::Input, Array = Self>;
+
+    fn builder() -> Self::Builder;
+}
+
+pub trait ArrayBuilder: Send + Sized {
+    type Element;
+    type Array;
+
+    #[must_use]
+    fn with_capacity(self, _capacity: usize) -> Self {
+        self
+    }
+
+    fn push(&mut self, value: Self::Element);
+
+    fn build(self) -> Self::Array;
 }