Builds on #852.

Adds support for vectorization of bitwise protocols, and adds vectorized version of `compare_gt` (used by quicksort).

Unvectorized bitwise protocols operate on `AdditiveShare<BA>`, as before (bits of the BA are different bits of multi-bit values associated with a particular record). Vectorized bitwise protocols operate on `BitDecomposed<AdditiveShare<BA>>`, where the elements of the outer `BitDecomposed` are different bits of multi-bit values associated with a particular record, and bits of the inner BA are the values of a particular bit position for a stripe of records. Adds an `ArrayAccessRef` variant of the `ArrayAccess` trait that is implemented for both kinds of inputs, and an `ArrayBuilder` trait that is implemented for both kinds of outputs.

## Performance

`compare_gt` benchmark. Generates 131,072 64-bit values and compares each of them to a single 64-bit threshold. This is a proxy for quicksort.

* Unvectorized: 39.6 s 🐢
* Vectorized x256, bitvec version (what is in this PR): 320 ms
* Vectorized x256, raw array version (see #928): 200 ms
* Vectorized x1024, raw array version: 100 ms
* Vectorized x4096, raw array version: 84 ms 🔥