@akoshelev 2024-01-24
> Vectorized x4096, raw array version: 84 ms 🔥

this is just awesome